# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transaction API",
        "description": "API for transaction control between users",
        "contact": {
            "email": "wendel.lopes777@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/transaction/{transaction}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Transaction view",
                "description": "Specific transaction view from id",
                "operationId": "transaction.view",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "description": "Transaction identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns transaction resource"
                    },
                    "404": {
                        "description": "Transaction not found response"
                    }
                }
            }
        },
        "/api/transaction": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Creating a transaction",
                "description": "Creation of a transaction informing a valid payer, beneficiary and value",
                "operationId": "transaction.create",
                "parameters": [
                    {
                        "name": "value",
                        "in": "query",
                        "description": "Transaction value",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "payer",
                        "in": "query",
                        "description": "User payer identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "payee",
                        "in": "query",
                        "description": "User payee identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns transaction resource"
                    },
                    "422": {
                        "description": "Request rules validation failed"
                    },
                    "500": {
                        "description": "Transaction creation error"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}